+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/paymentmethod/AddPaymentMethodRequirement.kt
@@ -1,6 +1,5 @@
 package com.stripe.android.lpmfoundations.paymentmethod
 
-import com.stripe.android.core.utils.FeatureFlags
 import com.stripe.android.model.LinkMode
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.PaymentMethod.Type.USBankAccount
@@ -81,9 +80,7 @@ private val PaymentMethodMetadata.supportsMobileInstantDebitsFlow: Boolean
         val paymentMethodTypes = stripeIntent.paymentMethodTypes
         val noUsBankAccount = USBankAccount.code !in paymentMethodTypes
         val supportsBankAccounts = "bank_account" in stripeIntent.linkFundingSources
-        val isDeferred = stripeIntent.clientSecret == null
-        return noUsBankAccount && supportsBankAccounts && canShowBankForm &&
-            (!isDeferred || FeatureFlags.instantDebitsDeferredIntent.isEnabled)
+        return noUsBankAccount && supportsBankAccounts && canShowBankForm
     }
 
 private val PaymentMethodMetadata.canShowBankForm: Boolean