+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/paymentdatacollection/ach/USBankAccountFormViewModel.kt
@@ -637,7 +637,7 @@ internal class USBankAccountFormViewModel @Inject internal constructor(
         }
 
         return PaymentSelection.New.USBankAccount(
-            labelResource = labelResource ?: "••••",
+            label = labelResource ?: "••••",
             iconResource = TransformToBankIcon(bankName),
             paymentMethodCreateParams = paymentMethodCreateParams,
             paymentMethodOptionsParams = paymentMethodOptionsParams,